+++ plugins/SebBinds/src/SebBinds/Properties/AssemblyInfo.cs
@@ -2,5 +2,5 @@ using System.Reflection;
 
 [assembly: AssemblyTitle("SebBinds")]
 [assembly: AssemblyProduct("SebBinds")]
-[assembly: AssemblyVersion("1.0.2")]
-[assembly: AssemblyFileVersion("1.0.2")]
+[assembly: AssemblyVersion("1.0.3")]
+[assembly: AssemblyFileVersion("1.0.3")]